+++ app/src/components/feature/overview/Overview.component.jsx
@@ -12,19 +12,19 @@ import { observer } from 'mobx-react';
 import PropTypes from 'prop-types';
 import { useContext, useEffect } from 'react';
 import { RootStoreContext } from 'stores/RootStore';
-import LineChart from '../stats/chart/LineChart.component';
-import WidgetContainer from '../stats/widgetcontainer/WidgetContainer';
-import SelectedComponentListComponent from '../stats/component/ComponentStats.component';
-import SelectedNodeListComponent from '../stats/node/NodeStats.component';
-import GraphStatsComponent from '../stats/graph/GraphStats.component';
-import ConnectionStatsComponent from '../stats/connections/ConnectionStats.component';
-import NodeFilterComponent from '../stats/nodefilter/NodeFilter.component';
+import LineChart from '../widgets/chart/LineChart.component';
+import WidgetContainer from '../widgets/widgetcontainer/WidgetContainer';
+import SelectedComponentListComponent from '../widgets/component/ComponentStats.component';
+import SelectedNodeListComponent from '../widgets/node/NodeStats.component';
+import GraphStatsComponent from '../widgets/graph/GraphStats.component';
+import ConnectionStatsComponent from '../widgets/connections/ConnectionStats.component';
+import NodeFilterComponent from '../widgets/nodefilter/NodeFilter.component';
 import { useState } from 'react';
 import { useResizeDetector } from 'react-resize-detector';
 import { useCallback } from 'react';
-import DoughnutChart from '../stats/chart/DoughnutChart.component';
-import BarChart from '../stats/chart/BarChart.component';
-import RadarChartComponent from '../stats/chart/RadarChart.component';
+import DoughnutChart from '../widgets/chart/DoughnutChart.component';
+import BarChart from '../widgets/chart/BarChart.component';
+import RadarChartComponent from '../widgets/chart/RadarChart.component';
 
 function Overview(props) {
     const store = useContext(RootStoreContext);
