+++ app/src/components/feature/widgets/chart/BarChart.component.jsx
