+++ app/src/components/feature/widgets/component/ComponentStats.component.jsx
