+++ app/src/components/feature/widgets/connections/ConnectionStats.component.jsx
