+++ app/src/components/feature/widgets/graph/GraphStats.component.jsx
